FromScratch/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
pillow>=7.1.2
matplotlib>=3.2.2
tqdm>=4.41.0
pandas>=1.1.4
scikit-learn>=0.24.1
albumentations>=1.0.3
tensorboard>=2.4.1
opencv-python>=4.5.1.48
pycocotools>=2.0.2 